requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
face-recognition>=1.3.0
dlib>=19.24.0
requests>=2.31.0

# Optional: For production
# gunicorn>=21.2.0
# celery>=5.3.0
# redis>=5.0.0
# psycopg2-binary>=2.9.0  # For PostgreSQL
